fixup tests for ppx_deriving.6.0.2,  ppx_deriving_yojson.3.8.0
The unit-tests now work for the latest versions of these two packages.
Since the latest changes to these packages have broken their proper
function when invoked via ocamlfind, the unit-tests needed to be
modified to account for that.  When invoked under ocamlfind, these
packages act as if "with_path = false" was passed as an option.  So a
constructor Foo in a file bar.ml is printed as "Foo"  and not as "Bar.Foo".